backend/requirements.txt
asgiref==3.2.7
autopep8==1.5.2
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.9.2
defusedxml==0.7.0rc1
Django==3.0.6
django-cors-headers==3.2.1
django-resized==0.3.11
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
drf-yasg==1.17.1
idna==2.9
inflection==0.4.0
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==20.3
Pillow==7.1.2
psycopg2-binary==2.8.5
pycodestyle==2.5.0
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
python3-openid==3.1.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
simplejson==3.17.0
six==1.14.0
social-auth-app-django==3.1.0
social-auth-core==3.3.3
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.9
